claude-for-msft-365-install/examples/python-bootstrap/requirements.txt
fastapi
uvicorn
PyJWT[crypto]
